more dmap

+++ trunk/pyNastran/bdf/dmap/spike.py
@@ -1,6 +1,7 @@
+# this is a comment
 a = 1
 b = 2
-c = 3.
+c = double(3.)
 d = 4.
 intval    = a + b
 floatval  = c + d
@@ -12,3 +13,10 @@ frogdogv = frogv + dogv
 
 #A = zeros((2,3),'f')
 #B = ones((2,3),'f')
+
+#f21 = open('bar_modes_lam.nas','wb')
+#f22 = open('bar_modes_mpf.nas','wb')
+#ASSIGN OUTPUT4=’bar_modes_lam.nas’,STATUS=unknown,UNIT=21
+
+#SOL 101
+#f21.write(a)
